Use uint32_t for count of jobs when querying job state
Continuation of commits 11da0b1357 and d404626940.
CID 349575.

Bug 18390

+++ src/common/slurm_protocol_defs.h
@@ -704,7 +704,7 @@ typedef struct job_info_request_msg {
 } job_info_request_msg_t;
 
 typedef struct {
-	uint16_t job_id_count;
+	uint32_t job_id_count;
 	uint32_t *job_ids;
 } job_state_request_msg_t;
